requirements.txt
django-extensions==3.1.3
pytest-pythonpath==0.7.3
pytest-django==4.4.0
django-enumfields==2.1.1
pytest==6.2.5
astral==2.2
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.10
environ==1.0
geographiclib==1.52
geopy==2.2.0
idna==3.2
numpy==1.21.2
psycopg2-binary==2.9.2
pytz==2021.1
requests==2.26.0
sqlparse==0.4.2
timezonefinder==5.2.0
urllib3==1.26.6
